Cleaned up some test cases
* Updated the tests to use the latest ways to
parse in the API
* Moved positive test cases to the top of the test case
files as this is what you first need to read to understand what the
feature is about.

Made most exception classes package private.

Added the possibility to declare a description inside of Command
classes. This should make it easier to produce nice usage texts while
still invoking CommandLineParser.forCommands(...) instead of
CommandLineParser.forArguments(...)

Added support for combined options such as "-fs" instead of "-f -s"

Removed validation of enum classes in the EnumParser constructor because
it caused the enum class to initialize.