Moved Finalizer/Finalizers into the internal package to minimize the public API.
Added validValuesDescription() in Limiter, this is now printed in the
usage if a limiter is used.

Removed CompoundLimiter as it wouldn't work properly with
RangeLimiterForRadix.

Moved parsing/printing logic in RadixiableParser into NumberType and
Radix so that it uses the strategy pattern instead.

Changed printing of numbers to remove leading zeros.

Change-Id: I01f545110fb30a0f2af3a6e47b44ee19e21ec0bc