Removed the possibility to set a radix explicitly.
The radix is now automatically detected by the leading characters of the
input instead. Binary input is no longer supported but I think a 400
line decrease in code base size makes up for that.

If someone only wants to support, for instance, decimal numbers they can
use ForwardingStringParser and implement this themselves.